+++ inst/scripts/mime-type.ps1
@@ -1,0 +1,7 @@
+# get the MIME type using .NET's MimeMapping class
+
+param (
+  [string]$filePath
+)
+
+Write-Host [System.Web.MimeMapping]::GetMimeMapping($filePath)